Split Atom into Atom and AtomicExpr, in conformity with the Basic/Expr split.
