+++ sympy/core/symbol.py
@@ -1,12 +1,13 @@
+from core import C
 from sympify import sympify
-from basic import Atom, S, C
-from expr import Expr
+from singleton import S
+from expr import Expr, AtomicExpr
 from cache import cacheit
 from sympy.logic.boolalg import Boolean
 
 import re
 
-class Symbol(Atom, Expr, Boolean):
+class Symbol(AtomicExpr, Boolean):
     """
     Assumptions::
        commutative = True
